+++ docs/changelog.md
@@ -9,6 +9,8 @@
 ## Версия 1.5.1
 - Новый модуль [Player](/func?id=player). Нужно [обновить права доступа](/install?id=Обновить-права-доступа).
 - `getPlayback` перемещен в `Player`.
+- `removeTracks` и `removeArtists` получили режим проверки только основного исполнителя трека.
+- Изменен формат входных параметров у [replaceWithSimilar](/func?id=replacewithsimilar).
 
 ## Версия 1.5.0
 - Учитывать всех исполнителей трека в фильтр-функциях `remove*`, `match*`, `dedup*`.